Frontend view model changes

+++ ExperimentalCMS.ViewModels/PlaceViewModel.cs
@@ -24,10 +24,10 @@ namespace ExperimentalCMS.ViewModels
         public int InternalRanking { get; set; }
 
         [Required]
-        public float Latitude { get; set; }
+        public decimal Latitude { get; set; }
 
         [Required]
-        public float Longitude { get; set; }
+        public decimal Longitude { get; set; }
 
         public int? PlaceTypeId { get; set; }
         public PlaceType PlaceType { get; set; }
